+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. The format 
 ## Unreleased
 ### Fixed
 - Accordion hover state now waits for pointer movement on Windows touchscreens
+- Accordion constrained height no longer oscillates when scrolling
 
 ## [v0.8.1]
 ### Improved